demo/pyproject.toml
[project]
name = "llm-learn"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain-anthropic>=0.3.12",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.31",
    "python-dotenv>=1.2.1",
]
